afl-fuzz: add unix socket mode
The current implementation of AFL is using stdin and files as the input
for the application. Introduce new unix socket mode for AFL:
  afl-fuzz ... -u <sock_name> ...
In this mode fuzzer will connect to the sock_name unix socket and all
the testcases will be send thought this socket. The socket can be the
fuzzed application itself or just a proxy which will redirect input to
the fuzzed application.

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>